Add a build script to construct a MeSH to PMID table
NCBI MEDLINE annotates articles (PMIDs) with MeSH terms.  I could
not find a bulk download of the PMID<->MeSH links, but the NCBI
eSearch tool provides a relatively easy way to obtain the links.
This script wraps eSearch to build a MeSH<->PMID table, which should
be useful in distant supervisio of phenotype mentions.